Moved selectParameterFile to deploy.ts
