+++ src/vscode-bicep/src/commands/deploy.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 import * as path from "path";
-import vscode from "vscode";
+import vscode, { Uri } from "vscode";
 import {
   LanguageClient,
   TextDocumentIdentifier,
@@ -10,11 +10,11 @@ import {
 import {
   AzExtTreeDataProvider,
   IActionContext,
+  IAzureQuickPickItem,
   parseError,
   UserCancelledError,
 } from "@microsoft/vscode-azext-utils";
 
-import { selectParameterFile } from "../deploy/selectParameterFile";
 import { AzLoginTreeItem } from "../deploy/tree/AzLoginTreeItem";
 import { AzResourceGroupTreeItem } from "../deploy/tree/AzResourceGroupTreeItem";
 import { LocationTreeItem } from "../deploy/tree/LocationTreeItem";
@@ -245,4 +245,56 @@ async function sendDeployCommand(
   );
 
   appendToOutputChannel(deploymentResponse);
+}
+
+async function selectParameterFile(
+  _context: IActionContext,
+  sourceUri: Uri | undefined
+): Promise<string> {
+  const quickPickItems: IAzureQuickPickItem[] =
+    await createParameterFileQuickPickList();
+  const result: IAzureQuickPickItem = await _context.ui.showQuickPick(
+    quickPickItems,
+    {
+      canPickMany: false,
+      placeHolder: `Select a parameter file`,
+      suppressPersistence: true,
+    }
+  );
+
+  if (result.label.includes("Browse...")) {
+    const paramsPaths: Uri[] | undefined = await vscode.window.showOpenDialog({
+      canSelectMany: false,
+      defaultUri: sourceUri,
+      openLabel: "Select Parameter File",
+    });
+    if (paramsPaths && paramsPaths.length == 1) {
+      const parameterFilePath = paramsPaths[0].fsPath;
+      appendToOutputChannel(
+        `Parameter file used in deployment -> ${path.basename(
+          parameterFilePath
+        )}`
+      );
+      return parameterFilePath;
+    }
+  }
+
+  appendToOutputChannel(`Parameter file was not provided`);
+
+  return "";
+}
+
+async function createParameterFileQuickPickList(): Promise<
+  IAzureQuickPickItem[]
+> {
+  const none: IAzureQuickPickItem = {
+    label: "$(circle-slash) None",
+    data: undefined,
+  };
+  const browse: IAzureQuickPickItem = {
+    label: "$(file-directory) Browse...",
+    data: undefined,
+  };
+
+  return [none].concat([browse]);
 }
